requirements.txt
python-dotenv==1.0.0
streamlit==1.22.0
requests~=2.30.0